requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
redis==5.0.1
google-generativeai==0.3.2
python-dotenv==1.0.1
python-multipart==0.0.7
gunicorn==21.2.0